Add a material pipeline to the renderer
The material pipeline makes it possible to register DocumentObjects with
an "enhanceMaterial" method, or a Proxy with an "enhanceMaterial".
This methods will be called in order of registration to customize the
RenderMaterial for the given object.

This makes the whole material handling pretty flexible as other workbenches
can implement this method and the user can add them to the material
pipeline of a render project.