Add a Camera object
This camera object allows to record a particular position of active Coin
camera and reuse it in rendering process.
Camera object can simply be inserted as a "view" in rendering projects,
like other rendered objects, and will thereby be exported to renderers.
Benefits for users are:
- User can easily access Camera settings for validation and tweaking.
- Camera object has a graphical representation that helps user see
where Camera is and where it looks to
- User can create as many Cameras as he/she wishes, to record various
points of views and consequently reuse them in rendering.
- User can reset active Coin camera back to any Camera object position,
in order to retrieve previous points of views
- User can opt to have no Camera in a given project: in that case, the
active Coin camera position will be used for rendering.

Camera object replaces the legacy handling of camera in rendering
projects as a (scarcely editable) project property. Camera property in
rendering projects is thus deprecated. However, a backward compatibility
function (Camera.retrieveLegacyCamera) is provided, which can create a
Camera object from the Camera field of a given rendering project.