Sunsky: Sun support in sunsky light for Cycles
This commit adds sun support in sunsky light. Cycles does not
provide any sun support in Hosek-Wilkie sky node, so this commit fills
the gap. It involves calculating the power and color of the sun for a given
tilt and atmospheric turbidity; these calculations have been grouped in a
separate module.