Materials: adaptations/simplifications to passthrough format
Passthrough required format was previously in python Format Specification
Mini-language, which was not very handy in INI file (FCMat).
In particular, one had to use double brackets ('{{' / '}}') which could be
disconcerting and error-prone for FCMat writer.
This previous is replaced with a new simpler one, with pseudovariables
(%NAME%, %RED%, %GREEN%, %BLUE%), which is later internally converted to
Python FSML.