Renderers: export into standard unit of distance (meter)
The internal unit of distance in FreeCAD is the millimeter. However, the
standard unit for renderers is the meter, to ensure consistency with
other SI units (watt, candela etc.), as well as with distances in other
data sources (textures, materials...).  Internal distances must
therefore be converted when exporting to the renderers.  This conversion
is a prerequisite for the implementation of texture features.