Projects: run renderer in a non-blocking mode (whatever renderer)
This commit adds the capability to run the renderer in an asynchronous
manner, whatever renderer it is, including CLI renderers.
In particular, renderer's stdout/stderr are piped to 'report view' in
FCD Gui.
At this stage, the code is in draft state, yet fully operational. There
is still lot of work to do in order to clean it and fix a few
non-critical issues.