Executor: use QThread
This commit is a proof of concept. Compared to Python native threads,
QThread is safer in Gui mode, and allows to trigger display only at the
end of rendering. However, it does not work in console mode, as no
QApplication runs in this mode.
Some work still remains to unify both approaches (native threads /
QThread).
Also, at this stage, code is not lintered.