get_rendering_string: don't use 'source.Proxy.type' attr to chose RenderHandler
Scripted objects can create/set arbitrary attributes in their Proxy
objects. Formerly, if an object's Proxy had a 'type' attribute, it would
be used to determine which Rendering Handler should be used. This behaviour
caused errors e.g. for FastenersWB objects, which used the type param
to indicate the specification of a fastener.

This commit changes the behaviour of the get_rendering_string function
to check if an object needs a specific handler using the Proxy's python type.